Other: Move selection post fixer to model/utils and inject it in model's constructor.
