+++ src/model/model.js
@@ -26,6 +26,7 @@ import insertContent from './utils/insertcontent';
 import deleteContent from './utils/deletecontent';
 import modifySelection from './utils/modifyselection';
 import getSelectedContent from './utils/getselectedcontent';
+import { injectSelectionPostFixer } from './utils/selection-postfixer';
 
 /**
  * Editor's data model. Read about the model in the
@@ -111,6 +112,8 @@ export default class Model {
 		this.schema.register( '$marker', {
 			allowIn: [ '$root', '$block' ]
 		} );
+
+		injectSelectionPostFixer( this );
 	}
 
 	/**
